+++ packages/rocketh-diamond/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rocketh/diamond
 
+## 0.11.17
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.11.15
+  - @rocketh/deploy@0.11.15
+  - @rocketh/read-execute@0.11.15
+
 ## 0.11.16
 
 ### Patch Changes