+++ packages/rocketh/CHANGELOG.md
@@ -1,5 +1,11 @@
 # rocketh
 
+## 0.11.15
+
+### Patch Changes
+
+- for convenience:support bigint but stringify it
+
 ## 0.11.14
 
 ### Patch Changes